+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - PR template
 - Test against 3.10
+- Compatibility with IPython Notebooks
 
 ## [0.5.0] - 2021-11-09
 ### Added